Add prediction and reflection to readme
The library has been vastly improved. Invoking commands is a one-liner.
Receiving functions have the values already unwrapped (read as a
NetIdMessage derivative). [MessageCommand] and [MessageRpc] functions
are tagged with an attribute and discovered using reflection.
Client-side prediction is built-in (although does nothing to reconcile
differences between server and client).