Fix bug causing client to jump back to server data
The lerping code incorporates a SyncVar, which can be clobbered by the
server's historical data. In that case, the client observes the lerp
suddenly jerk back towards a previous state.

The solution was to send SyncVars only once each to each joining client.
This way, the SyncVar guarantees all clients join the game with the
correct state, but then can operate on their local version without being
clobbered by the server.