Fix error on joining mid-command/rpc
If a client received a message callback after connecting, but before
instantiating the receiving game object, an error would be thrown.

The method LocalAuthorityBehaviour.FindComponent() has been changed to
return null if the object or component does not exist. Now, callers are
responsible for null checking.

Added a null-check exit early condition in the callback lambdas. This is
correct behaviour. Despite missing the callback, the objects are still
instantiated with the correct state. The proper state is received
through the OnSerialize/OnSerialize methods.